+++ ImageFeed/ImagesList/ImagesListViewController.swift
@@ -24,11 +24,12 @@ class ImagesListViewController: UIViewController {
         tableView.delegate = self
         tableView.dataSource = self
         guard let token = tokenStorage.token else {return}
+        //        tableView.reloadData()
         imagesListService.fetchPhotosNextPage(token: token)
         UIBlockingProgressHUD.show()
         imagesListServiceObserver = NotificationCenter.default
             .addObserver(
-                forName: ImageListService.DidChangeNotification,
+                forName: ImageListService.didChangeNotification,
                 object: nil,
                 queue: .main
             ) { [weak self] _ in
@@ -36,6 +37,11 @@ class ImagesListViewController: UIViewController {
                 UIBlockingProgressHUD.dismiss()
                 self.updateTableViewAnimated()
             }
+    }
+
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+        tableView.reloadData()
     }
     
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
@@ -58,18 +64,22 @@ class ImagesListViewController: UIViewController {
         willDisplay cell: UITableViewCell,
         forRowAt indexPath: IndexPath
     ) {
-        guard let token = tokenStorage.token else {return}
-        if (indexPath.row + 1 == photos.count){
-            imagesListService.fetchPhotosNextPage(token: token)
-        }
+        guard let token = tokenStorage.token, indexPath.row + 1 == photos.count else {return}
+        imagesListService.fetchPhotosNextPage(token: token)
     }
+
     func configCell(for cell: ImagesListCell, with indexPath: IndexPath)  {
-        cell.cellImage.image = UIImage(named: "placeholder")
         let url = URL(string: photos[indexPath.row].thumbImageURL)
-        let processor = RoundCornerImageProcessor(cornerRadius: 35)
+        let processor = RoundCornerImageProcessor(cornerRadius: 16)
         cell.cellImage.kf.indicatorType = .activity
-        cell.cellImage.kf.setImage(with: url, options: [.processor(processor)])
-        tableView.reloadRows(at: [indexPath], with: .automatic)
+        cell.cellImage.kf.setImage(with: url, placeholder: UIImage(named: "placeholder"), options: [.processor(processor)]){ [self] result in
+            switch result {
+            case .success:
+                self.tableView.reloadRows(at: [indexPath], with: .automatic)
+            case .failure(let error):
+                print(error)
+            }
+        }
         if let date = photos[indexPath.row].createdAt {
             cell.dateLabel.text = dateFormatter.string(from: date)
         }
@@ -94,14 +104,15 @@ class ImagesListViewController: UIViewController {
         let oldCount = photos.count
         let newCount = imagesListService.photos.count
         photos = imagesListService.photos
-        if oldCount != newCount {
-            tableView.performBatchUpdates {
-                let indexPaths = (oldCount..<newCount).map { i in
-                    IndexPath(row: i, section: 0)
-                }
-                tableView.insertRows(at: indexPaths, with: .automatic)
-            } completion: { _ in }
-        }
+        guard oldCount != newCount else {return}
+        //        if oldCount != newCount {
+        tableView.performBatchUpdates {
+            let indexPaths = (oldCount..<newCount).map { i in
+                IndexPath(row: i, section: 0)
+            }
+            tableView.insertRows(at: indexPaths, with: .automatic)
+        } completion: { _ in }
+        //        }
     }
 }
 
@@ -121,5 +132,6 @@ extension ImagesListViewController: ImagesListCellDelegate {
         guard let token = tokenStorage.token else {return}
         UIBlockingProgressHUD.show()
         imagesListService.changeLike(photoId: photo.id, isLike: !photo.isLiked, token: token, cell.setIsLiked)
+        UIBlockingProgressHUD.dismiss()
     }
 }
