+++ ImageFeed/Models/LikedPhotoResult.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+struct LikedPhotoResult: Codable {
+    let photo: PhotoResult
+}
