+++ ImageFeed/Models/Photo.swift
@@ -1,0 +1,11 @@
+import UIKit
+
+struct Photo {
+    let id: String
+    let size: CGSize
+    let createdAt: Date?
+    let welcomeDescription: String?
+    let thumbImageURL: String
+    let largeImageURL: String
+    let isLiked: Bool
+}
