+++ ImageFeed/Models/Profile.swift
@@ -1,0 +1,8 @@
+import UIKit
+
+struct Profile {
+    let username: String
+    let name: String
+    let loginName : String
+    let bio: String?
+}
