+++ ImageFeed/Models/ProfileImage.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+struct ProfileImage: Codable {
+    let small: String?
+    let medium: String?
+    let large: String?
+}
