+++ ImageFeed/Models/UrlsResult.swift
@@ -1,0 +1,9 @@
+import UIKit
+
+struct UrlsResult: Codable {
+    let raw: String
+    let full: String
+    let regular: String
+    let small: String
+    let thumb: String
+}
