+++ ImageFeed/Models/UserResult.swift
@@ -1,0 +1,9 @@
+import UIKit
+
+struct UserResult: Codable {
+    let profileImage: ProfileImage?
+
+    enum CodingKeys: String, CodingKey {
+        case profileImage = "profile_image"
+    }
+}
